refactor : 객체간(Lotto, WinningNumber) 상호작용을 통해 로또등수 반환하도록 수정
